chore: Separate AppWIndow into two types
