+++ src/Uno.UWP/Microsoft/UI/IconId.cs
@@ -2,30 +2,29 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Microsoft.UI
+namespace Microsoft.UI;
+
+public struct IconId : IEquatable<IconId>
 {
-	public struct IconId : IEquatable<IconId>
+	public ulong Value;
+
+	public IconId(ulong _Value)
 	{
-		public ulong Value;
-
-		public IconId(ulong _Value)
-		{
-			Value = _Value;
-		}
-
-		public static bool operator ==(IconId x, IconId y)
-			=> x.Value == y.Value;
-
-		public static bool operator !=(IconId x, IconId y)
-			=> x.Value != y.Value;
-
-		public bool Equals(IconId other)
-			=> Value == other.Value;
-
-		public override bool Equals(object obj)
-			=> (obj is IconId y) && Equals(y);
-
-		public override int GetHashCode()
-			=> Value.GetHashCode();
+		Value = _Value;
 	}
+
+	public static bool operator ==(IconId x, IconId y)
+		=> x.Value == y.Value;
+
+	public static bool operator !=(IconId x, IconId y)
+		=> x.Value != y.Value;
+
+	public bool Equals(IconId other)
+		=> Value == other.Value;
+
+	public override bool Equals(object obj)
+		=> (obj is IconId y) && Equals(y);
+
+	public override int GetHashCode()
+		=> Value.GetHashCode();
 }
