+++ src/Uno.UWP/UI/ViewManagement/ApplicationView.cs
@@ -14,12 +14,7 @@ using Windows.Foundation;
 using Windows.Storage;
 using Windows.UI.WindowManagement;
 using MUXWindowId = Microsoft.UI.WindowId;
-
-#if HAS_UNO_WINUI
 using AppWindow = Microsoft.UI.Windowing.AppWindow;
-#else
-using AppWindow = Windows.UI.WindowManagement.AppWindow;
-#endif
 
 namespace Windows.UI.ViewManagement
 {
