--- src/Uno.UWP/UI/Windowing/AppWindow.Uwp.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-#if HAS_UNO_WINUI
-namespace Microsoft.UI.Windowing;
-#else
-namespace Windows.UI.WindowManagement;
-#endif
-
-partial class AppWindow
-{
-}
